refactor(sanitize): extract sanitizer core to internal/sanitize
Move the pure, filesystem-free sanitization logic out of the cmd/niac
main package into an importable internal/sanitize package (Mapping,
Options, Stats, Content). The deterministic sha256 IP mapping is
preserved byte-for-byte. cmd/niac keeps cobra wiring, path-allowlist
validation, and file I/O, now calling internal/sanitize; the .tmp write
is confined via os.Root so path traversal is impossible by construction.
Add library.ReadFile as the os.Root-contained read counterpart to
WriteFile. Pre-alpha: callers migrated in-place, no compat shim.
